Move common functions and variable to base class
