+++ Source/UEActionGame/Private/Characters/BaseCharacter.cpp
@@ -3,6 +3,7 @@
 
 #include "Characters/BaseCharacter.h"
 #include "Components/BoxComponent.h"
+#include "Components/AttributeComponent.h"
 #include "Items/Weapons/Weapon.h"
 
 // Sets default values
@@ -11,6 +12,8 @@ ABaseCharacter::ABaseCharacter()
  	// Set this character to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
 	PrimaryActorTick.bCanEverTick = true;
 
+	Attributes = CreateDefaultSubobject<UAttributeComponent>(TEXT("Attributes"));
+
 }
 
 // Called when the game starts or when spawned
@@ -18,6 +21,72 @@ void ABaseCharacter::BeginPlay()
 {
 	Super::BeginPlay();
 	
+}
+
+void ABaseCharacter::Attack()
+{
+}
+
+void ABaseCharacter::PlayAttackMontage()
+{
+}
+
+bool ABaseCharacter::CanAttack()
+{
+	return false;
+}
+
+void ABaseCharacter::Die()
+{
+}
+
+void ABaseCharacter::AttackEnd()
+{
+}
+
+void ABaseCharacter::GetHit(const FVector& ImpactPoint)
+{
+}
+
+void ABaseCharacter::PlayHitReactMontage(const FName SectionName)
+{
+	UAnimInstance* AnimInstance = this->GetMesh()->GetAnimInstance();
+	if (AnimInstance && HitReactMontage)
+	{
+		AnimInstance->Montage_Play(HitReactMontage);
+		AnimInstance->Montage_JumpToSection(SectionName, HitReactMontage);
+	}
+}
+
+void ABaseCharacter::DirectionalHitReact(const FVector& ImpactPoint)
+{
+	const FVector Forward = GetActorForwardVector();
+	const FVector ImpactLowered(ImpactPoint.X, ImpactPoint.Y, GetActorLocation().Z);
+	const FVector ToHit = (ImpactLowered - GetActorLocation()).GetSafeNormal();
+
+	const double CosTheta = FVector::DotProduct(Forward, ToHit);
+	double Theta = FMath::Acos(CosTheta);
+	Theta = FMath::RadiansToDegrees(Theta);
+
+	// if CrossProduct points down, Theta should be negative
+	const FVector CrossProduct = FVector::CrossProduct(Forward, ToHit);
+	if (CrossProduct.Z < 0)
+	{
+		Theta *= -1.f;
+		//UKismetSystemLibrary::DrawDebugArrow(this, GetActorLocation(), GetActorLocation() + CrossProduct * 100.f, 5.f,
+		//	FColor::Blue, 5.f);
+	}
+
+	FName SectionName("ReactBack"); // Because 135 > Theta < -135 means needs to react to the back
+
+	if (Theta >= -45.f && Theta < 45.f)
+		SectionName = FName("ReactFront");
+	else if (Theta >= -135 && Theta < 45)
+		SectionName = FName("ReactLeft");
+	else if (Theta >= 45 && Theta < 135)
+		SectionName = FName("ReactRight");
+
+	this->PlayHitReactMontage(SectionName);
 }
 
 // Called every frame
