+++ Source/UEActionGame/Public/Characters/BaseCharacter.h
@@ -4,12 +4,15 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/Character.h"
+#include "Interfaces/HitInterface.h"
 #include "BaseCharacter.generated.h"
 
 class AWeapon;
+class UAttributeComponent;
+class UAnimMontage;
 
 UCLASS()
-class UEACTIONGAME_API ABaseCharacter : public ACharacter
+class UEACTIONGAME_API ABaseCharacter : public ACharacter, public IHitInterface
 {
 	GENERATED_BODY()
 
@@ -23,7 +26,41 @@ public:
 protected:
 	virtual void BeginPlay() override;
 
+	virtual void Attack();
+
+	virtual void PlayAttackMontage();
+
+	void PlayHitReactMontage(const FName SectionName);
+
+	void DirectionalHitReact(const FVector& ImpactPoint);
+
+	virtual bool CanAttack();
+
+	virtual void Die();
+
+	UFUNCTION(BlueprintCallable)
+	virtual	void AttackEnd();
+
+	virtual void GetHit(const FVector& ImpactPoint) override;
+
 	UPROPERTY(VisibleAnywhere, Category = Weapon)
 	AWeapon* EquippedWeapon;
+
+	UPROPERTY(EditDefaultsOnly, Category = Montages)
+	UAnimMontage* AttackMontage;
+
+	UPROPERTY(EditDefaultsOnly, Category = Animation)
+	UAnimMontage* HitReactMontage;
+
+	UPROPERTY(EditDefaultsOnly, Category = Animation)
+	UAnimMontage* DeathMontage;
+
+	UPROPERTY(VisibleAnywhere)
+	UAttributeComponent* Attributes;
+
+	UPROPERTY(EditAnywhere, Category = Effects)
+	USoundBase* HitSound;
+	UPROPERTY(EditAnywhere, Category = Effects)
+	UParticleSystem* HitParticles;
 
 };
