+++ Source/UEActionGame/Public/Interfaces/HitInterface.h
@@ -21,6 +21,6 @@ class UEACTIONGAME_API IHitInterface
 	GENERATED_BODY()
 
 	// Add interface functions to this class. This is the class that will be inherited to implement this interface.
-public:
+public: 
 	virtual void GetHit(const FVector& ImpactPoint) = 0;
 };
